Updated jennkins clean stage to be conditional based on pipeline progression

+++ Jenkinsfile
@@ -3,22 +3,40 @@ pipeline {
     agent any
 
     environment {
-        image_label = "ab-users-microservice"
-        git_commit_hash = "${sh(returnStdout: true, script: 'git rev-parse --short=8 HEAD')}"
-        image = ""
+        image_label = "users-microservice-ab"
+        commit = sh(returnStdout: true, script: "git rev-parse --short=8 HEAD").trim()
+        image = null
+        packaged = false
+        built = false
     }
 
     stages {
         stage('Package') {
             steps {
-                sh "./mvnw clean package"
+                sh "./mvnw package"
+            }
+
+            post {
+                success {
+                    script {
+                        packaged = true
+                    }
+                }
             }
         }
 
         stage('Build') {
             steps {
                 script {
-                    image = docker.build image_label
+                    image = docker.build(image_label)
+                }
+            }
+
+            post {
+                success {
+                    script {
+                        built = true
+                    }
                 }
             }
         }
@@ -26,18 +44,30 @@ pipeline {
         stage('Push to registry') {
             steps {
                 script {
-                    docker.withRegistry(USERS_ECR_URI_AB, 'ecr:us-west-2:ecr-creds') {
-                        image.push("$git_commit_hash")
+                    ecr_repo_uri ="https://${ORG_ACCOUNT_NUM}.dkr.ecr.${region}.amazonaws.com/${image_label}"
+                    docker.withRegistry(ecr_repo_uri, "ecr:$region:ecr-creds") {
+                        image.push("$commit")
                         image.push('latest')
                     }
                 }
             }
         }
+    }
 
-        stage('Clean up') {
-            steps {
-                sh "docker rmi $image_label"
+    post {
+        cleanup {
+            script {
+                if(packaged) {
+                    sh "./mvnw clean"
+
+                    if(built) {
+                        sh "docker rmi $image_label"
+                    }
+                }
             }
         }
     }
 }
+
+
+
